Bugfix: sort globbed files
Consider something like this in the config file:

    extension jpg
    extension tif

When working with a directory that contains files of both types the
files previously wouldn't get sorted by file name but by extension,
apparently because of how file globbing with a pattern like
`*.{jpg,tif}` works. This commit fixes this, sorting files like
expected.